fix: cleanup fp command

+++ configs/zsh/functions/fp.zsh
@@ -2,33 +2,26 @@
 
 function fp(){
 
-    for p in "$@"
-    do
+    if [ "$(which lsof)" = "lsof not found" ]; then
+        echo "lsof not installed"
+    else
+        for PORT in "$@"
+        do
+           if [ "$PORT" != "" ]; then
+                #echo "checking ($PORT)"
+                PIDS="$(lsof -t -i:$PORT)"
+                AMOUNT=$(wc -w<<<$PIDS);
 
-        PORT=$var;
-
-        if [ "$(which lsof)" = "lsof not found" ]; then
-            echo "lsof not installed"
-        else
-            PIDS="$(lsof -t -i:$PORT)"
-
-            AMOUNT=$(wc -w<<<$PIDS);
-
-            if [ $AMOUNT = 0 ]; then
-
-                echo "No processes use port $PORT";
-
-            else
-
-                echo "Killing $AMOUNT process(es) using port $PORT";
-
-                echo "$PIDS" | while IFS= read -r pid ; do 
-                    kill -9 $pid;
-                done
-
+                if [ $AMOUNT = 0 ]; then
+                    echo "No processes use port $PORT";
+                else
+                    echo "Killing $AMOUNT process(es) using port $PORT";
+                    echo "$PIDS" | while IFS= read -r pid ; do 
+                        kill -9 $pid;
+                    done
+                fi
             fi
-        fi
-
-    done
+        done
+    fi
 
 }
